Fix double to string conversion

+++ Arduino15/packages/SPRESENSE/hardware/spresense/1.0.0/cores/spresense/WString.cpp
@@ -54,7 +54,7 @@ static char* utoa(int value, char * buf, int base)
     return buf;
 }
 
-static char * dtostrf(float value, unsigned int width, unsigned int decimalPlaces, char* buf)
+static char * dtostrf(double value, unsigned int width, unsigned int decimalPlaces, char* buf)
 {
     char fmt[20];
